Stringutils: just StripPrefix and Canonicalize
I'd like a bit more consistency in our generation, so we'll have a "canonical
representation" for all prefix tokens. Note that this ISN'T what we actually
output: it's only useful as a way to key the map, so that "Together" and
"together" map to the same suffix keys.

The stringutils files provide tools to modify or alter the tweets during the
input phase, but hopefully also the output phase, if we decide to go that far.