Operates on "canonical" prefixes
Previously, we'd map "It's dark" and "its dark" to different suffixes,
now they are stored in canonical form "its dark", and we branch from there.
Note that we don't lose the representational data (the difference between
"its" and "It's") in the output, since we store them in a separate map.

Might want to have this configurable though, since I'm noticing a dip
in the quality of the tweets it's generating...