Datastore now working for Generators and Tweets
Now we compare fetched tweets with what we store in the datastore, and store
Generators persistently.

Still missing the "deep dive" (where we can fetch as many tweets as we can for
a user), which limits our corpus, as well as top-level Google Account
integration. And, of course, prettifying this thing.