Fixed one datastore bug, introduce another.
So we know actually read and write to the datastore, unlike before when
I either didn't check for errors, or it failed opaquely (hint: 'noindex'
will kill an entire query, even if you didn't search by that attribute).

Now we only fetch ~36/39 tweets even though we store hundreds or
thousands. Also, dev_appserver can only store 1000? Really? Not
documented anywhere, but seems to be the experience...