Some improvements to text generation, TODO adds
In trying to make the text generation better, rather than pick random
prefixes from absolutely anywhere in the prefix map, we now delineate
"beginnings" in the Generators, which take the first prefix of all data
sources. This isn't wholly accurate (we'd rather take the first prefix
of each sentence, so a call to AddSeeds with a blog post wouldn't only
pay attention to the first prefix of the whole post) but it's a start.