First desugaring of AppEngine.
AppEngine, while pretty cool in many ways, has also been a collossal pain
in the ass, namely the datastore. It's behavior is wonky (only 1000 records?)
and the code/test/debug cycle is slow, and would require a ton more tools
to get right (namely: no unit tests, having to alt-tab to browser + refresh to
see if code had the desired effect, dependence on SDK, syncronous code [fetching
tweets] holds up the rest of the render, etc...)

Unless data storage and retrieval is happening correctly (and IT HASN'T) I
can't really focus on a myriad of other improvements, or ever launching this
thing. I expect to go back to App Engine eventually, but for the time being, I'm
removing it and just creating a simple service that can be accessed with the
command-line.

This commit removes app engine calls, replaces them with to-be-implemented stubs.
It builds correctly, though I haven't run my old tests.