Seed random-number generator, fix DB bug.
The old random-number generator was deterministic, so it would produce the
same tweets in order. We now seed it before running with the system clock.

Also fix a DB bug where we keep re-inserting the newest tweets we can into
the DB. The issue arose from the fact that we can't store 64-bit ints into
SQLite (Twitter Id's are predictably huge) so we have to store them as strings.
Relying on SQLite's ORDER BY clause meant that it was sorting strings, so the
"top" result was usually the shortest string, by digits. We now do the sorting
in the Go layer.