Representations now are flaggable.
You can choose to let the Markov bot work on "canonical" versions of data
(so "Its'/It's/ITS/its" would all map to the same) or with the raw strings.
Before I'd added the "canonical" data as a feature, but it was mandatory.
Turning it off, however, leads to much better nonsense tweets.